Refactor directory structure.
I moved away from the vague name 'infra' and to a more
descriptive name 'gateways'. The new name makes it easier to
see that the modules therein are entrypoints to disk or to
the internet.

I also added a few helper modules to the 'application' folder
and moved the document patcher to the 'application' folder.
The document patcher is a pure class that does no file I/O
so it doesn't make sense for it to live in 'gateways'.